+++ pkg/api/routes.go
@@ -1,14 +1,22 @@
 package api
 
 import (
+	"context"
+	"encoding/base64"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 	"github.com/hydrolix/clickhouse-sql-parser/parser"
 	"github.com/hydrolix/plugin/pkg/datasource"
+	"github.com/hydrolix/plugin/pkg/models"
 	"maps"
 	"net/http"
+	"net/http/httputil"
+	"net/url"
 	"slices"
+	"strings"
 	"time"
 )
 
@@ -131,14 +139,75 @@ func wrapError(rw http.ResponseWriter, err error) {
 	return
 }
 
-func Routes(ds *datasource.HydrolixDatasource) map[string]func(http.ResponseWriter, *http.Request) {
+func NewAssistantProxyHandler(routePrefix string, settings models.PluginSettings) (http.HandlerFunc, error) {
+	errorHandler := func(rw http.ResponseWriter, req *http.Request, err error) {
+		status := http.StatusBadGateway
+		if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+			status = http.StatusRequestTimeout
+		}
+
+		log.DefaultLogger.Error(
+			"Assistant proxy error",
+			"method", req.Method,
+			"url", req.URL.String(),
+			"status", status,
+			"error", err,
+		)
+
+		response, _ := json.Marshal(map[string]string{"detail": err.Error()})
+
+		rw.Header().Set("Content-Type", "application/json")
+		rw.WriteHeader(status)
+		_, _ = rw.Write(response)
+	}
+	aiBaseUrl := settings.AiBaseUrl
+	if aiBaseUrl == nil {
+		defaultUrl, err := url.Parse(fmt.Sprintf("https://%s:4040/assistant", settings.Host))
+		if err != nil {
+			return nil, err
+		}
+		aiBaseUrl = defaultUrl
+	}
+
+	proxy := httputil.NewSingleHostReverseProxy(aiBaseUrl)
+	proxy.ErrorHandler = errorHandler
+
+	director := proxy.Director
+	proxy.Director = func(req *http.Request) {
+		director(req)
+		req.URL.Path = strings.TrimPrefix(req.URL.Path, routePrefix)
+		if settings.CredentialsType == "serviceAccount" {
+			req.Header.Set("Authorization", "Bearer "+settings.Token)
+		} else {
+			auth := settings.UserName + ":" + settings.Password
+			auth = base64.StdEncoding.EncodeToString([]byte(auth))
+			req.Header.Set("Authorization", "Basic "+auth)
+		}
+	}
+
+	return proxy.ServeHTTP, nil
+}
+
+func Routes(ds *datasource.HydrolixDatasource, ctx context.Context, settings backend.DataSourceInstanceSettings) (map[string]func(http.ResponseWriter, *http.Request), error) {
+	pluginSettings, err := models.NewPluginSettings(ctx, settings)
+	if err != nil {
+		return nil, err
+	}
+
+	assistantPrefix := "/assistant"
+	assistantHandler, err := NewAssistantProxyHandler(assistantPrefix, pluginSettings)
+	if err != nil {
+		return nil, err
+	}
+
 	return map[string]func(http.ResponseWriter, *http.Request){
 		"/ast": AST,
 		"/interpolate": func(writer http.ResponseWriter, request *http.Request) {
 			Interpolate(ds, writer, request)
 		},
-		"/macroCTE": MacroCTEs,
-	}
+		"/macroCTE":           MacroCTEs,
+		assistantPrefix + "/": assistantHandler,
+	}, nil
 }
 
 type Request[T any] struct {