+++ src/components/ConfigEditor.tsx
@@ -154,6 +154,7 @@ export function ConfigEditor(props: Props) {
       },
     });
   };
+
   const onProtocolToggle = (protocol: Protocol) => {
     onOptionsChange({
       ...options,
@@ -233,6 +234,33 @@ export function ConfigEditor(props: Props) {
       },
     });
   };
+
+  const onAiEnabledChange = (value: boolean) => {
+    let jsonData = {
+      ...options.jsonData,
+      aiEnabled: value,
+    };
+    if (value && !jsonData.aiBaseUrl) {
+      jsonData["useDefaultAiBaseUrl"] = true;
+      jsonData["aiBaseUrl"] = `https://${jsonData.host}:4040/assistant`;
+    }
+    onOptionsChange({
+      ...options,
+      jsonData,
+    });
+  };
+
+  const onUseDefaultAiBaseUrlChange = (useDefault: boolean) => {
+    onOptionsChange({
+      ...options,
+      jsonData: {
+        ...options.jsonData,
+        useDefaultAiBaseUrl: useDefault,
+        aiBaseUrl: `https://${jsonData.host}:4040/assistant`,
+      },
+    });
+  };
+
   const onQuerySettingsChange = (key: string) => {
     return (value: string) => {
       let querySettings = jsonData?.querySettings ?? [];
@@ -659,6 +687,57 @@ export function ConfigEditor(props: Props) {
               type="number"
             />
           </Field>
+          <Divider />
+          <ConfigSection title="Assistant">
+            <Field
+              data-testid={labels.aiEnabled.testId}
+              label={labels.aiEnabled.label}
+              description={labels.aiEnabled.description}
+            >
+              <Switch
+                name={"aiEnabled"}
+                className="gf-form"
+                value={jsonData.aiEnabled}
+                onChange={(e) => onAiEnabledChange(e.currentTarget.checked)}
+              />
+            </Field>
+            {jsonData.aiEnabled && (
+              <div>
+                <Field
+                  data-testid={labels.aiBaseUrl.testId}
+                  label={labels.aiBaseUrl.label}
+                  description={labels.aiBaseUrl.description}
+                >
+                  <Stack direction="row">
+                    <Input
+                      name={"aiBaseUrl"}
+                      width={40}
+                      value={jsonData.aiBaseUrl || ""}
+                      onChange={onUpdateDatasourceJsonDataOption(
+                        props,
+                        "aiBaseUrl"
+                      )}
+                      disabled={jsonData.useDefaultAiBaseUrl}
+                      label={labels.aiBaseUrl.label}
+                      aria-label={labels.aiBaseUrl.label}
+                    />
+                    <InlineField
+                      data-testId={labels.useDefaultAiBaseUrl.testId}
+                      label={labels.useDefaultAiBaseUrl.label}
+                      interactive
+                    >
+                      <InlineSwitch
+                        onChange={(e) =>
+                          onUseDefaultAiBaseUrlChange(e.currentTarget.checked)
+                        }
+                        value={jsonData.useDefaultAiBaseUrl}
+                      />
+                    </InlineField>
+                  </Stack>
+                </Field>
+              </div>
+            )}
+          </ConfigSection>
           <Divider />
           <ConfigSection title="Query Settings">
             <div style={{ marginBottom: "20px" }}>